Add customization callbacks to LLVMTypeConverter.
This implementation is a stepping stone towards a pattern-based type
conversion approach that allows customization of different points of
the type conversion and different types for each customization point.

Re current implementation:
- It only allows customization of different points but not customization
of different types for each customziation point for now but it should
be easy to extend once the latter is needed.
- LLVMTypeConverterCustomization holds the callbacks for the customizations
points available (only function argument types for now) and it's
initialized by default to the pre-existing calling convention.
- 'makeStandardToLLVMBarePtrTypeConverter' is private on purpose to
avoid a combinatorial explosion similar to that of populate* functions.
Customization methods are made public to allow external clients to
create their ad-hoc customizations.